+++ src/Symfony/Bridge/PhpUnit/DeprecationErrorHandler.php
@@ -123,6 +123,9 @@ class DeprecationErrorHandler
         }
 
         $deprecation = new Deprecation($msg, debug_backtrace(), $file);
+        if ($deprecation->isMuted()) {
+            return;
+        }
         $group = 'other';
 
         if ($deprecation->originatesFromAnObject()) {