+++ src/Symfony/Bridge/PhpUnit/DeprecationErrorHandler/Deprecation.php
@@ -48,9 +48,9 @@ class Deprecation
     private $originMethod;
 
     /**
-     * @var string one of the PATH_TYPE_* constants
+     * @var string
      */
-    private $triggeringFilePathType;
+    private $triggeringFile;
 
     /** @var string[] absolute paths to vendor directories */
     private static $vendors;
@@ -76,7 +76,7 @@ class Deprecation
             // No-op
         }
         $line = $trace[$i];
-        $this->triggeringFilePathType = $this->getPathType($file);
+        $this->triggeringFile = $file;
         if (isset($line['object']) || isset($line['class'])) {
             if (isset($line['class']) && 0 === strpos($line['class'], SymfonyTestsListenerFor::class)) {
                 $parsedMsg = unserialize($this->message);
@@ -88,7 +88,7 @@ class Deprecation
                 // then we need to use the serialized information to determine
                 // if the error has been triggered from vendor code.
                 if (isset($parsedMsg['triggering_file'])) {
-                    $this->triggeringFilePathType = $this->getPathType($parsedMsg['triggering_file']);
+                    $this->triggeringFile = $parsedMsg['triggering_file'];
                 }
 
                 return;
@@ -170,6 +170,21 @@ class Deprecation
     }
 
     /**
+     * @return bool
+     */
+    public function isMuted()
+    {
+        if ('Function ReflectionType::__toString() is deprecated' !== $this->message) {
+            return false;
+        }
+        if (isset($this->trace[1]['class'])) {
+            return 0 === strpos($this->trace[1]['class'], 'PHPUnit\\');
+        }
+
+        return false !== strpos($this->triggeringFile, \DIRECTORY_SEPARATOR.'vendor'.\DIRECTORY_SEPARATOR.'phpunit'.\DIRECTORY_SEPARATOR);
+    }
+
+    /**
      * Tells whether both the calling package and the called package are vendor
      * packages.
      *
@@ -177,10 +192,11 @@ class Deprecation
      */
     public function getType()
     {
-        if (self::PATH_TYPE_SELF === $this->triggeringFilePathType) {
+        $triggeringFilePathType = $this->getPathType($this->triggeringFile);
+        if (self::PATH_TYPE_SELF === $triggeringFilePathType) {
             return self::TYPE_SELF;
         }
-        if (self::PATH_TYPE_UNDETERMINED === $this->triggeringFilePathType) {
+        if (self::PATH_TYPE_UNDETERMINED === $triggeringFilePathType) {
             return self::TYPE_UNDETERMINED;
         }
         $erroringFile = $erroringPackage = null;