+++ src/Symfony/Component/Lock/BlockingStoreInterface.php
@@ -12,7 +12,6 @@
 namespace Symfony\Component\Lock;
 
 use Symfony\Component\Lock\Exception\LockConflictedException;
-use Symfony\Component\Lock\Exception\NotSupportedException;
 
 /**
  * @author Hamza Amrouche <hamza.simperfit@gmail.com>
@@ -22,15 +21,7 @@ interface BlockingStoreInterface
     /**
      * Waits until a key becomes free, then stores the resource.
      *
-     * If the store does not support this feature it should throw a NotSupportedException.
-     *
      * @throws LockConflictedException
-     * @throws NotSupportedException
      */
     public function waitAndSave(Key $key);
-
-    /**
-     * Checks if the store can wait until a key becomes free before storing the resource.
-     */
-    public function supportsWaitAndSave(): bool;
 }