+++ src/Symfony/Component/Lock/Lock.php
@@ -37,12 +37,12 @@ final class Lock implements LockInterface, LoggerAwareInterface
     private $dirty = false;
 
     /**
-     * @param Key                   $key         Resource to lock
-     * @param PersistStoreInterface $store       Store used to handle lock persistence
-     * @param float|null            $ttl         Maximum expected lock duration in seconds
-     * @param bool                  $autoRelease Whether to automatically release the lock or not when the lock instance is destroyed
+     * @param Key                      $key         Resource to lock
+     * @param PersistingStoreInterface $store       Store used to handle lock persistence
+     * @param float|null               $ttl         Maximum expected lock duration in seconds
+     * @param bool                     $autoRelease Whether to automatically release the lock or not when the lock instance is destroyed
      */
-    public function __construct(Key $key, PersistStoreInterface $store, float $ttl = null, bool $autoRelease = true)
+    public function __construct(Key $key, PersistingStoreInterface $store, float $ttl = null, bool $autoRelease = true)
     {
         $this->store = $store;
         $this->key = $key;
@@ -71,7 +71,7 @@ final class Lock implements LockInterface, LoggerAwareInterface
     {
         try {
             if ($blocking) {
-                if (!$this->store instanceof StoreInterface && !($this->store instanceof BlockingStoreInterface && $this->store->supportsWaitAndSave())) {
+                if (!$this->store instanceof StoreInterface && !$this->store instanceof BlockingStoreInterface) {
                     throw new NotSupportedException(sprintf('The store "%s" does not support blocking locks.', \get_class($this->store)));
                 }
                 $this->store->waitAndSave($this->key);