+++ src/Symfony/Component/Lock/Store/MemcachedStore.php
@@ -71,10 +71,12 @@ class MemcachedStore implements StoreInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated since Symfony 4.4.
      */
     public function waitAndSave(Key $key)
     {
-        @trigger_error(sprintf('%s() is deprecated since Symfony 4.4 and will be removed in Symfony 5.0.', __METHOD__));
+        @trigger_error(sprintf('%s() is deprecated since Symfony 4.4 and will be removed in Symfony 5.0.', __METHOD__), E_USER_DEPRECATED);
         throw new InvalidArgumentException(sprintf('The store "%s" does not supports blocking locks.', \get_class($this)));
     }