+++ src/Symfony/Component/Lock/Tests/Store/CombinedStoreTest.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\Lock\Tests\Store;
 use Symfony\Component\Lock\BlockingStoreInterface;
 use Symfony\Component\Lock\Exception\LockConflictedException;
 use Symfony\Component\Lock\Key;
-use Symfony\Component\Lock\PersistStoreInterface;
+use Symfony\Component\Lock\PersistingStoreInterface;
 use Symfony\Component\Lock\Store\CombinedStore;
 use Symfony\Component\Lock\Store\RedisStore;
 use Symfony\Component\Lock\Strategy\StrategyInterface;
@@ -62,8 +62,8 @@ class CombinedStoreTest extends AbstractStoreTest
     protected function setUp()
     {
         $this->strategy = $this->getMockBuilder(StrategyInterface::class)->getMock();
-        $this->store1 = $this->getMockBuilder([PersistStoreInterface::class, BlockingStoreInterface::class])->getMock();
-        $this->store2 = $this->getMockBuilder([PersistStoreInterface::class, BlockingStoreInterface::class])->getMock();
+        $this->store1 = $this->getMockBuilder([PersistingStoreInterface::class, BlockingStoreInterface::class])->getMock();
+        $this->store2 = $this->getMockBuilder([PersistingStoreInterface::class, BlockingStoreInterface::class])->getMock();
 
         $this->store = new CombinedStore([$this->store1, $this->store2], $this->strategy);
     }
@@ -267,8 +267,8 @@ class CombinedStoreTest extends AbstractStoreTest
 
     public function testPutOffExpirationIgnoreNonExpiringStorage()
     {
-        $store1 = $this->getMockBuilder(PersistStoreInterface::class)->getMock();
-        $store2 = $this->getMockBuilder(PersistStoreInterface::class)->getMock();
+        $store1 = $this->getMockBuilder(PersistingStoreInterface::class)->getMock();
+        $store2 = $this->getMockBuilder(PersistingStoreInterface::class)->getMock();
 
         $store = new CombinedStore([$store1, $store2], $this->strategy);